Suggested changes including other tests
