+++ tests/x.py
@@ -1,0 +1,1 @@
+RawRequestMessage(method='GET', path='/', version=HttpVersion(major=1, minor=1), headers=<CIMultiDict('Host': 'localhost:8080', 'Connection': 'keep-alive', 'Upgrade-Insecure-Requests': '1', 'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.119 Safari/537.36', 'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8', 'Accept-Encoding': 'gzip, deflate, br', 'Accept-Language': 'en-US,en;q=0.9,ta;q=0.8', 'Cookie': '_ga=GA1.1.1724746481.1517250860; tz=Asia%2FCalcutta; __utma=111872281.1724746481.1517250860.1517278865.1517278865.1; __utmc=111872281; __utmz=111872281.1517278865.1.1.utmcsr=(direct)|utmccn=(direct)|utmcmd=(none); optimizelyEndUserId=oeu1517280742880r0.7999913885797751; optimizelySegments=%7B%22201523968%22%3A%22false%22%2C%22201527938%22%3A%22none%22%2C%22201527939%22%3A%22gc%22%2C%22201574652%22%3A%22direct%22%7D; optimizelyBuckets=%7B%7D; utag_main=v_id:016144facff600119c9ca794027504069001a0610091a$_sn:6$_ss:0$_st:1518236380578$_pn:2%3Bexp-session$ses_id:1518234578691%3Bexp-session; sess_uuid=0417fbd1-3928-4568-bc4e-2ebac4e607b0')>, raw_headers=[(bytearray(b'HOST'), bytearray(b'localhost:8080')), (bytearray(b'CONNECTION'), bytearray(b'keep-alive')), (bytearray(b'UPGRADE-INSECURE-REQUESTS'), bytearray(b'1')), (bytearray(b'USER-AGENT'), bytearray(b'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/64.0.3282.119 Safari/537.36')), (bytearray(b'ACCEPT'), bytearray(b'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,image/apng,*/*;q=0.8')), (bytearray(b'ACCEPT-ENCODING'), bytearray(b'gzip, deflate, br')), (bytearray(b'ACCEPT-LANGUAGE'), bytearray(b'en-US,en;q=0.9,ta;q=0.8')), (bytearray(b'COOKIE'), bytearray(b'_ga=GA1.1.1724746481.1517250860; tz=Asia%2FCalcutta; __utma=111872281.1724746481.1517250860.1517278865.1517278865.1; __utmc=111872281; __utmz=111872281.1517278865.1.1.utmcsr=(direct)|utmccn=(direct)|utmcmd=(none); optimizelyEndUserId=oeu1517280742880r0.7999913885797751; optimizelySegments=%7B%22201523968%22%3A%22false%22%2C%22201527938%22%3A%22none%22%2C%22201527939%22%3A%22gc%22%2C%22201574652%22%3A%22direct%22%7D; optimizelyBuckets=%7B%7D; utag_main=v_id:016144facff600119c9ca794027504069001a0610091a$_sn:6$_ss:0$_st:1518236380578$_pn:2%3Bexp-session$ses_id:1518234578691%3Bexp-session; sess_uuid=0417fbd1-3928-4568-bc4e-2ebac4e607b0'))], should_close=False, compression=None)
